lib/pyproject.toml
[tool.poetry]
name = "hass_ae"
version = "0.1.0"
description = ""
authors = ["Mikael Hernrup <mikael@hernrup.se>"]

[tool.poetry.dependencies]
python = "^3.7"
asyncws = "^0.1"
websockets = "^8.1"
cachetools = "^4.1.1"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
hass_ae = "hass_ae.app:run"